+++ catalyst/contrib/dl/callbacks/kornia_transform.py
@@ -15,6 +15,79 @@ class BatchTransformCallback(Callback):
     Look at `Kornia: an Open Source Differentiable Computer Vision
     Library for PyTorch`_ for details.
 
+    Usage example for notebook API:
+
+    .. code-block:: python
+
+        import os
+
+        from kornia import augmentation
+        import torch
+        from torch.nn import functional as F
+        from torch.utils.data import DataLoader
+
+        from catalyst import dl
+        from catalyst.contrib.data.transforms import ToTensor
+        from catalyst.contrib.datasets import MNIST
+        from catalyst.contrib.dl.callbacks.kornia_transform import (
+            BatchTransformCallback
+        )
+        from catalyst.utils import metrics
+
+
+        class CustomRunner(dl.Runner):
+            def predict_batch(self, batch):
+                # model inference step
+                return self.model(
+                    batch[0].to(self.device).view(batch[0].size(0), -1)
+                )
+
+            def _handle_batch(self, batch):
+                # model train/valid step
+                x, y = batch
+                y_hat = self.model(x.view(x.size(0), -1))
+
+                loss = F.cross_entropy(y_hat, y)
+                accuracy01, *_ = metrics.accuracy(y_hat, y)
+                self.batch_metrics.update(
+                    {"loss": loss, "accuracy01": accuracy01}
+                )
+
+                if self.is_train_loader:
+                    loss.backward()
+                    self.optimizer.step()
+                    self.optimizer.zero_grad()
+
+        model = torch.nn.Linear(28 * 28, 10)
+        optimizer = torch.optim.Adam(model.parameters(), lr=0.02)
+
+        loaders = {
+            "train": DataLoader(
+                MNIST(os.getcwd(), train=True, transform=ToTensor()),
+                batch_size=32,
+            ),
+            "valid": DataLoader(
+                MNIST(os.getcwd(), train=False, transform=ToTensor()),
+                batch_size=32,
+            ),
+        }
+        transrorms = [
+            augmentation.RandomAffine(degrees=(-15, 20), scale=(0.75, 1.25)),
+        ]
+
+        runner = CustomRunner()
+
+        # model training
+        runner.train(
+            model=model,
+            optimizer=optimizer,
+            loaders=loaders,
+            logdir="./logs",
+            num_epochs=5,
+            verbose=True,
+            callbacks=[BatchTransformCallback(transrorms, input_key=0)],
+        )
+
     To apply augmentations only during specific loader e.g. only during
     training :class:`catalyst.core.callbacks.control_flow.ControlFlowCallback`
     callback can be used. For config API it can look like this:
@@ -49,9 +122,9 @@ class BatchTransformCallback(Callback):
     def __init__(
         self,
         transform: Sequence[Union[dict, AugmentationBase]],
-        input_key: str = "image",
+        input_key: Union[str, int] = "image",
         additional_input_key: Optional[str] = None,
-        output_key: Optional[str] = None,
+        output_key: Optional[Union[str, int]] = None,
         additional_output_key: Optional[str] = None,
     ) -> None:
         """Constructor method for the :class:`BatchTransformCallback` callback.
@@ -69,10 +142,11 @@ class BatchTransformCallback(Callback):
                 an augmentation name as a value. Please note that in this case
                 to use custom augmentation you should add it to the
                 `TRANSFORMS` registry first.
-            input_key (str): key in batch dict mapping to transform,
-                e.g. `'image'`
-            additional_input_key (Optional[str]): key of an additional target
-                in batch dict mapping to transform, e.g. `'mask'`
+            input_key (Union[str, int]): key in batch dict
+                mapping to transform, e.g. `'image'`
+            additional_input_key (Optional[Union[str, int]]): key of an
+                additional target in batch dict mapping to transform,
+                e.g. `'mask'`
             output_key (Optional[str]): key to use to store the result
                 of the transform, defaults to `input_key` if not provided
             additional_output_key (Optional[str]): key to use to store
@@ -118,15 +192,16 @@ class BatchTransformCallback(Callback):
         return input_[self.input_key], input_[self.additional_input]
 
     def _process_output_tensor(
-        self, batch: Tuple[torch.Tensor, torch.Tensor]
+        self, runner: IRunner, batch: Tuple[torch.Tensor, torch.Tensor]
     ) -> Dict[str, torch.Tensor]:
-        return {self.output_key: batch}
+        runner.input[self.output_key] = batch
 
     def _process_output_tuple(
-        self, batch: Tuple[torch.Tensor, torch.Tensor]
-    ) -> Dict[str, torch.Tensor]:
+        self, runner: IRunner, batch: Tuple[torch.Tensor, torch.Tensor]
+    ) -> None:
         out_t, additional_t = batch
-        return {self.output_key: out_t, self.additional_output: additional_t}
+        dict_ = {self.output_key: out_t, self.additional_output: additional_t}
+        runner.input.update(dict_)
 
     def on_batch_start(self, runner: IRunner) -> None:
         """Apply transforms.
@@ -136,7 +211,7 @@ class BatchTransformCallback(Callback):
         """
         in_batch = self._process_input(runner.input)
         out_batch = self.transform(in_batch)
-        runner.input.update(self._process_output(out_batch))
+        self._process_output(runner, out_batch)
 
 
 __all__ = ["BatchTransformCallback"]